Cleanup pass 4: const-correct TrajectoryProtein + safe auto-fixes + test names
clang-tidy: 225 → 168 findings (-57). Cumulative: 3171 → 168 (94.7%
reduction from bootstrap baseline at 76e7068).

Three classes of fix landed in this pass:

1. Const-correct GetDenseBuffer (29 const_cast sites eliminated)

   TrajectoryProtein::GetDenseBuffer<T> gained a const overload
   returning const DenseBuffer<T>*. DenseBuffer<T> already had const
   overloads for AtomSlicePtr / RawData, so this completes the
   const-correctness chain through to the H5 emit path.

   Per-call-site change: `const_cast<TrajectoryProtein&>(tp).GetDense
   Buffer<...>(...)` → `tp.GetDenseBuffer<...>(...)` — same call, just
   on the const path. 29 TR WriteH5Group sites updated mechanically
   via grep+sed.

   Two const_cast remain — both legitimate patterns NOLINT'd inline
   with rationale:

   - src/ChargeAssignmentResult.cpp:27 — lazy-init pattern. Compute
     receives const conf (calculator base-class contract) but Protein
     lazy-builds ForceFieldChargeTable on first need. The mutation is
     cache-fill, not observable state change. Refactoring is a
     substrate-architecture decision, not a lint fix.

   - tests/test_rmsd_spike_dft_coord.cpp:250 — test verifies double-
     Finalize is idempotent (finalized_ guard). Needs non-const TR
     reference; tp.Result<T>() returns const ref by public contract.
     Test-only access pattern, not production.

2. Mass auto-fix on safe categories (28 files)

   run-clang-tidy -fix with the residual safe-checks set:
   - readability-static-definition-in-anonymous-namespace (13 sites)
   - readability-redundant-casting (9 sites)
   - readability-convert-member-functions-to-static (11 sites)
   - modernize-concat-nested-namespaces (5 sites)
   - misc-unused-using-decls (2 sites)
   - readability-qualified-auto (2 sites)
   - modernize-make-unique (1 site)
   Total ~43 fixes.

3. Gtest underscore-in-name renames (13 sites)

   google-readability-avoid-underscore-in-googletest-name fires when
   a test name contains '_'. Per Googletest FAQ, underscores in
   test/fixture names can collide with TEST_F macro expansion. 13
   tests renamed:
     AmberResidueThreeLetter_Canonical → AmberResidueThreeLetterCanonical
     AmberResidueThreeLetter_Variants → AmberResidueThreeLetterVariants
     IupacResidueThreeLetter_AlwaysCanonical → IupacResidueThreeLetterAlwaysCanonical
     ResidueOneLetter_Standard20 → ResidueOneLetterStandard20
     BmrbAtomName_AlaSidechain → BmrbAtomNameAlaSidechain
     BmrbStereoLabel_GlyHaPair → BmrbStereoLabelGlyHaPair
     BmrbStereoLabel_ArgMethyleneIsProR_Or_ProS → BmrbStereoLabelArgMethyleneIsProROrProS
     FixturesPathsResolveToRealFiles_1P9J → FixturesPathsResolveToRealFiles1P9J
     FixturesPathsResolveToRealFiles_1Z9B → FixturesPathsResolveToRealFiles1Z9B
     FullSystemReaderReadsAmberTpr_1P9J → FullSystemReaderReadsAmberTpr1P9J
     FullSystemReaderBuildsAmberProtein_1P9J → FullSystemReaderBuildsAmberProtein1P9J
     FullSystemReaderBuildsAmberProtein_1Z9B → FullSystemReaderBuildsAmberProtein1Z9B
     ProbeHNCAHAvsPhiOffset_1UBQ → ProbeHNCAHAvsPhiOffset1UBQ

Tests: 21/21 in focused TR / RmsdTracking / RmsdSpike / DftPoseCoord /
BsWelford / HmWelford / McWelford regression. Build green.

Top remaining categories (168 total):
  28  readability-function-size           NOLINT'd per calculator
  13  cert-err34-c                        real parse-error class
  10  bugprone-branch-clone               case-by-case (intentional vs merge)
   9  performance-inefficient-string-concatenation  perf
   8  concurrency-mt-unsafe               wrapper-NOLINT
   6  clang-diagnostic-unused-but-set-variable  real
   6  clang-diagnostic-unused-variable    real
   6  bugprone-implicit-widening-of-multiplication-result  real (cast)
   6  misc-unused-parameters              review
   6  misc-const-correctness              auto-fix straggler
   5  cert-flp30-c                        loop induction FP
   5  bugprone-unchecked-optional-access  false positives (refactor)
   5  bugprone-argument-comment           review
   5  cert-env33-c                        NOLINT (controlled inputs)
  ... long tail

cppcheck: 363 → 330 (-33; const-correctness fixes bled over).

+++ src/BsWelfordTrajectoryResult.cpp
@@ -262,7 +262,7 @@ void BsWelfordTrajectoryResult::WriteH5Group(
     grp.createAttribute("result_name",     Name());
     grp.createAttribute("n_frames",        n_frames_);
     grp.createAttribute("finalized",       finalized_);
-    grp.createAttribute("ddof",            static_cast<int>(1));
+    grp.createAttribute("ddof",            1);
     grp.createAttribute("mean_dt_ps",      mean_dt_ps_);
     grp.createAttribute("frame_index_range", frame_index_range_);
     // T1 is stored as Cartesian Levi-Civita dual (T1[0]=v_x, T1[1]=v_y,
