+++ src/HydrationGeometryWelfordTrajectoryResult.cpp
@@ -235,7 +235,7 @@ void HydrationGeometryWelfordTrajectoryResult::WriteH5Group(
     grp.createAttribute("n_frames",               n_frames_);
     grp.createAttribute("source_attached_count",  source_attached_count);
     grp.createAttribute("finalized",              finalized_);
-    grp.createAttribute("ddof",                   static_cast<int>(1));
+    grp.createAttribute("ddof",                   1);
     grp.createAttribute("mean_dt_ps",             mean_dt_ps_);
     grp.createAttribute("frame_index_range",      frame_index_range_);
     grp.createAttribute("irrep_layout_dipole",    std::string("v_x,v_y,v_z"));
