+++ src/McConnellWelfordTrajectoryResult.cpp
@@ -241,7 +241,7 @@ void McConnellWelfordTrajectoryResult::WriteH5Group(
     grp.createAttribute("result_name",     Name());
     grp.createAttribute("n_frames",        n_frames_);
     grp.createAttribute("finalized",       finalized_);
-    grp.createAttribute("ddof",            static_cast<int>(1));
+    grp.createAttribute("ddof",            1);
     grp.createAttribute("mean_dt_ps",      mean_dt_ps_);
     grp.createAttribute("frame_index_range", frame_index_range_);
     grp.createAttribute("irrep_layout_t1", std::string("v_x,v_y,v_z"));
